Fix endpoint delete issue

+++ pkg/controller/controller.go
@@ -79,7 +79,10 @@ func (c *controller) handleAddEndpoint(obj interface{}) {
 		utilruntime.HandleError(fmt.Errorf("unexpected object type: %v", obj))
 		return
 	}
-	c.proxy.AddEndpoints(ep)
+	if ep.Namespace == "default" && ep.Name == "portal" {
+		klog.Infof("Adding endpoint: %s/%s subsets: %+v", ep.Namespace, ep.Name, ep.Subsets)
+		c.proxy.AddEndpoints(ep)
+	}
 }
 
 func (c *controller) handleUpdateEndpoint(oldObj, newObj interface{}) {
@@ -94,10 +97,12 @@ func (c *controller) handleUpdateEndpoint(oldObj, newObj interface{}) {
 		return
 	}
 	if epOld.ObjectMeta.ResourceVersion == epNew.ObjectMeta.ResourceVersion {
-		klog.Infof("Resync update for service: %s/%s", epNew.ObjectMeta.Namespace, epNew.ObjectMeta.Name)
 		return
 	}
-	c.proxy.UpdateEndpoints(epOld, epNew)
+	if epNew.Namespace == "default" && epNew.Name == "portal" {
+		klog.Infof("Calling c.proxy.UpdateEndpoints for %s/%s subsets: %+v", epNew.Namespace, epNew.Name, epNew.Subsets)
+		c.proxy.UpdateEndpoints(epOld, epNew)
+	}
 }
 
 func (c *controller) handleDeleteEndpoint(obj interface{}) {
@@ -113,7 +118,10 @@ func (c *controller) handleDeleteEndpoint(obj interface{}) {
 			return
 		}
 	}
-	c.proxy.DeleteEndpoints(ep)
+	if ep.Namespace == "default" && ep.Name == "portal" {
+		klog.Infof("Delete endpoint: %s/%s subsets: %+v", ep.Namespace, ep.Name, ep.Subsets)
+		c.proxy.DeleteEndpoints(ep)
+	}
 }
 
 func (c *controller) Run(stopCh <-chan struct{}) error {
